refactor: Move everything Stripe related to Stripe module
