+++ apps/api/src/stripe/stripe-metadata.interface.ts
@@ -1,7 +1,7 @@
 import { DonationType } from '@prisma/client'
 import Stripe from 'stripe'
 
-export interface DonationMetadata extends Stripe.MetadataParam {
+export interface StripeMetadata extends Stripe.MetadataParam {
   type: DonationType | null
   campaignId: string | null
   personId: string | null
