+++ apps/api/src/stripe/stripe.module.ts
@@ -6,7 +6,9 @@ import { ConfigService } from '@nestjs/config'
 import { StripeConfigFactory } from '../donations/helpers/stripe-config-factory'
 import { CampaignModule } from '../campaign/campaign.module'
 import { PersonModule } from '../person/person.module'
-import { PrismaService } from '../prisma/prisma.service'
+
+import { DonationsModule } from '../donations/donations.module'
+import { RecurringDonationModule } from '../recurring-donation/recurring-donation.module'
 
 @Module({
   imports: [
@@ -16,6 +18,8 @@ import { PrismaService } from '../prisma/prisma.service'
     }),
     CampaignModule,
     PersonModule,
+    DonationsModule,
+    RecurringDonationModule,
   ],
   providers: [StripeService],
   controllers: [StripeController],
